+++ src/daemon/command_parser_executor.h
@@ -36,8 +36,6 @@
 
 #pragma once
 
-#include <optional>
-
 #include "daemon/rpc_command_executor.h"
 #include "common/common_fwd.h"
 #include "rpc/core_rpc_server.h"
@@ -52,8 +50,6 @@ public:
   t_command_parser_executor(
       uint32_t ip
     , uint16_t port
-    , const std::optional<tools::login>& login
-    , const epee::net_utils::ssl_options_t& ssl_options
     , bool is_rpc
     , cryptonote::core_rpc_server* rpc_server = NULL
     );